refactor(network): reorganize network folder
