+++ src/network/p2p/node.test.ts
@@ -1,9 +1,9 @@
 import { Multiaddr } from '@multiformats/multiaddr/';
 import { ServerError } from '~/utils/errors';
 import { Factories } from '~/test/factories';
-import { Node } from '~/network/node';
-import { GossipMessage, NETWORK_TOPIC_PRIMARY } from '~/network/protocol';
-import { sleep } from '~/utils/utils';
+import { Node } from '~/network/p2p/node';
+import { GossipMessage, NETWORK_TOPIC_PRIMARY } from '~/network/p2p/protocol';
+import { sleep } from '~/utils/crypto';
 
 const NUM_NODES = 10;
