+++ src/network/rpc/index.ts
@@ -1,0 +1,7 @@
+/* eslint-disable no-restricted-imports */
+
+import { RPCClient } from './client';
+import { RPCServer } from './server';
+import { RPCHandler } from './utils';
+
+export { RPCClient, RPCServer, RPCHandler };
