+++ src/network/rpc/rpc.test.ts
@@ -20,7 +20,7 @@ import {
   VerificationRemove,
 } from '~/types';
 import { Factories } from '~/test/factories';
-import { generateEd25519Signer, generateEthereumSigner } from '~/utils/utils';
+import { generateEd25519Signer, generateEthereumSigner } from '~/utils/crypto';
 import { RPCServer, RPCClient, RPCHandler } from '~/network/rpc';
 import Engine from '~/storage/engine';
 import Faker from 'faker';
