+++ src/storage/db/verification.test.ts
@@ -2,7 +2,7 @@ import { Factories } from '~/test/factories';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import { Ed25519Signer, VerificationEthereumAddress, VerificationRemove } from '~/types';
 import { NotFoundError } from '~/utils/errors';
-import { generateEd25519Signer } from '~/utils/utils';
+import { generateEd25519Signer } from '~/utils/crypto';
 import VerificationDB from '~/storage/db/verification';
 
 const rocks = jestRocksDB('db.verification.test');
