+++ src/storage/engine/mock.ts
@@ -1,6 +1,6 @@
 import { Ed25519Signer, EthereumSigner } from '~/types';
 import { Factories } from '~/test/factories';
-import { generateEd25519Signer, generateEthereumSigner } from '~/utils/utils';
+import { generateEd25519Signer, generateEthereumSigner } from '~/utils/crypto';
 import Engine from '~/storage/engine';
 import Faker from 'faker';
