+++ src/storage/sets/followSet.test.ts
@@ -2,7 +2,7 @@ import { Factories } from '~/test/factories';
 import Faker from 'faker';
 import FollowSet from '~/storage/sets/followSet';
 import { Ed25519Signer, Follow, FollowAdd, FollowRemove, URI } from '~/types';
-import { generateEd25519Signer } from '~/utils/utils';
+import { generateEd25519Signer } from '~/utils/crypto';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import FollowDB from '~/storage/db/follow';
 import { BadRequestError, NotFoundError } from '~/utils/errors';
