+++ src/storage/sets/followSet.ts
@@ -4,7 +4,7 @@ import RocksDB from '~/storage/db/rocksdb';
 import { BadRequestError } from '~/utils/errors';
 import { Follow, FollowAdd, FollowRemove, URI } from '~/types';
 import { isFollowAdd, isFollowRemove } from '~/types/typeguards';
-import { hashCompare } from '~/utils/utils';
+import { hashCompare } from '~/utils/crypto';
 
 /**
  * FollowSet is a modified LWW set that stores and fetches follow actions. FollowAdd and FollowRemove messages
