+++ src/storage/sets/reactionSet.ts
@@ -4,7 +4,7 @@ import RocksDB from '~/storage/db/rocksdb';
 import { BadRequestError } from '~/utils/errors';
 import { Reaction, ReactionAdd, ReactionRemove, URI } from '~/types';
 import { isReactionAdd, isReactionRemove } from '~/types/typeguards';
-import { hashCompare } from '~/utils/utils';
+import { hashCompare } from '~/utils/crypto';
 
 /**
  * ReactionSet is a modified LWW set that stores and fetches reactions. ReactionAdd and ReactionRemove messages are
