+++ src/test/e2e/hub.test.ts
@@ -4,8 +4,8 @@ import { generateUserInfo, getIDRegistryEvent, getSignerAdd, mockFid, populateEn
 import { Factories } from '~/test/factories';
 import { Hub, HubOpts } from '~/hub';
 import { RPCClient } from '~/network/rpc';
-import { sleep } from '~/utils/utils';
-import { ContactInfoContent, Content, GossipMessage, NETWORK_TOPIC_PRIMARY } from '~/network/protocol';
+import { sleep } from '~/utils/crypto';
+import { ContactInfoContent, Content, GossipMessage, NETWORK_TOPIC_PRIMARY } from '~/network/p2p/protocol';
 import { Message } from '~/types';
 import { ServerError } from '~/utils/errors';
 import { jest } from '@jest/globals';
